requirement.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.26.post2
celery-with-redis==3.0
Django==1.11.11
djangorestframework==3.7.7
kombu==3.0.37
pytz==2018.3
redis==2.10.6
vine==1.1.4
